pyproject.toml
[tool.poetry]
name = "claude-gran-cassa"
version = "0.1.0"
description = "Claude AI powered techno"
authors = ["Paul Evans <paulevans@theyard.studio>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
anthropic = "^0.42.0"
python-dotenv = "^1.0.1"
midiutil = "^1.2.1"
pygame = "^2.6.1"
mido = "^1.3.3"
click = "^8.1.8"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
claude-gran-cassa = "claude_gran_cassa.cli.main:main"
